[map] refactor(map-plugin) remove redundant code
