+++ packages/plugin-map/src/MapViewer.jsx
@@ -1,13 +1,45 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { validate } from 'wix-rich-content-common';
+import { withGoogleMap, GoogleMap, Marker, InfoWindow } from 'react-google-maps';
+import ReactGoogleMapLoader from 'react-google-maps-loader';
 import isEqual from 'lodash/isEqual';
+import { validate } from 'wix-rich-content-common';
 import schema from '../statics/data-schema.json';
-import { Map } from './Map';
+
+const GoogleMapWrapper = withGoogleMap(props => (
+  <GoogleMap
+    defaultCenter={{ lat: props.lat, lng: props.lng }}
+    center={{ lat: props.lat, lng: props.lng }}
+    zoom={props.zoom}
+    options={{
+      draggable: props.isDraggingAllowed,
+      mapTypeId: props.mode,
+      zoomControl: props.isZoomControlShown,
+      streetViewControl: props.isStreetViewControlShown,
+      mapTypeControl: false,
+    }}
+  >
+    <Marker
+      options={{ visible: props.isMarkerShown }}
+      title={props.markerTitle}
+      position={{ lat: props.lat, lng: props.lng }}
+      onClick={props.onMarkerClick}
+    >
+      {props.isMarkerTooltipRendered && (
+        <InfoWindow onCloseClick={props.onMarkerTooltipCloseClick}>
+          <div>{props.markerTooltipContent}</div>
+        </InfoWindow>
+      )}
+    </Marker>
+  </GoogleMap>
+));
 
 export class MapViewer extends Component {
   constructor(props) {
     super(props);
+    this.state = {
+      isMarkerTooltipRendered: true,
+    };
     validate(props.componentData, schema);
   }
 
@@ -18,23 +50,44 @@ export class MapViewer extends Component {
   }
 
   render() {
-    return <Map {...this.props} />;
+    const { componentData } = this.props;
+
+    return (
+      <ReactGoogleMapLoader
+        params={{
+          key: componentData.googleMapApiKey,
+          libraries: 'geometry,drawing,places',
+        }}
+        render={googleMaps =>
+          googleMaps && (
+            <GoogleMapWrapper
+              isMarkerShown={componentData.mapSettings.isMarkerShown}
+              loadingElement={<div style={{ height: `100%` }} />}
+              containerElement={<div style={{ height: componentData.config.height || '400px' }} />}
+              mapElement={<div style={{ height: `100%` }} />}
+              lat={Number(componentData.mapSettings.lat)}
+              lng={Number(componentData.mapSettings.lng)}
+              markerTitle={componentData.mapSettings.address}
+              markerTooltipContent={componentData.mapSettings.locationDisplayName}
+              zoom={componentData.mapSettings.zoom}
+              onMarkerTooltipCloseClick={() => this.setState({ isMarkerTooltipRendered: false })}
+              onMarkerClick={() =>
+                this.setState({ isMarkerTooltipRendered: !this.state.isMarkerTooltipRendered })
+              }
+              isMarkerTooltipRendered={this.state.isMarkerTooltipRendered}
+              mode={componentData.mapSettings.mode}
+              isZoomControlShown={componentData.mapSettings.isZoomControlShown}
+              isStreetViewControlShown={componentData.mapSettings.isStreetViewControlShown}
+              isDraggingAllowed={componentData.mapSettings.isDraggingAllowed}
+              {...this.props}
+            />
+          )
+        }
+      />
+    );
   }
 }
 
 MapViewer.propTypes = {
   componentData: PropTypes.object.isRequired,
-  theme: PropTypes.object,
-  onReady: PropTypes.func,
-  onStart: PropTypes.func,
-  controls: PropTypes.bool,
-  width: PropTypes.string,
-  height: PropTypes.string,
-  settings: PropTypes.object,
-};
-
-MapViewer.defaultProps = {
-  width: '100%',
-  height: '100%',
-  controls: true,
 };
